refactor: common 패키지 위치 이동
common 패키지를 domain 이하 폴더로 이동.
domain에 공통적으로 사용되는 것이므로 domain 이하가 적당하다고 판단.

Issue #29
