+++ src/main/java/com/hhplus/commerce/domain/common/exception/CommonErrorCode.java
@@ -1,4 +1,4 @@
-package com.hhplus.commerce.common.exception;
+package com.hhplus.commerce.domain.common.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
